requirements.txt
yt-dlp==2024.3.10
PyYAML==6.0.1
